Git报告生成器/Git报告生成器.app/Contents/Resources/requirements.txt
PyQt6>=6.4.0
    GitPython>=3.1.30
    pandas>=1.5.3
    jinja2>=3.1.2
    python-dateutil>=2.8.2
    Pillow>=9.0.0
    